requirements.txt
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
kaggle==2.0.0
kagglesdk==0.1.16
packaging==26.0
protobuf==7.34.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
requests==2.32.5
six==1.17.0
text-unidecode==1.3
tqdm==4.67.3
urllib3==2.6.3
webencodings==0.5.1
